requirements.txt
fastapi[standard]
python-dotenv
pydantic[email,timezone]
oracledb
pytest
sqlalchemy
alembic
passlib[bcrypt] pyjwt